+++ lesson-14/src/App.jsx
@@ -1,0 +1,53 @@
+// rrd imports
+import { createBrowserRouter, Navigate, RouterProvider } from "react-router";
+
+// layouts
+import MainLayout from "./layouts/MainLayout";
+
+// pages
+import Home from "./pages/Home";
+import Login from "./pages/Login";
+import Signup from "./pages/Signup";
+
+// components
+import ProtectedRoutes from "./components/ProtectedRoutes";
+
+function App() {
+  const user = false;
+  const routes = createBrowserRouter([
+    {
+      path: "/",
+      element: (
+        <ProtectedRoutes user={user}>
+          <MainLayout />
+        </ProtectedRoutes>
+      ),
+      children: [
+        {
+          index: true,
+          element: <Home />,
+        },
+        {
+          path: "/about",
+          element: <h1>About</h1>,
+        },
+        {
+          path: "/contact",
+          element: <h1>Contact</h1>,
+        },
+      ],
+    },
+    {
+      path: "/login",
+      element: user ? <Navigate to="/" /> : <Login />,
+    },
+    {
+      path: "/signup",
+      element: user ? <Navigate to="/" /> : <Signup />,
+    },
+  ]);
+
+  return <RouterProvider router={routes} />;
+}
+
+export default App;
